orb_backend_status_types_redux

+++ backend-status/src/backend/types.rs
@@ -49,7 +49,7 @@ pub struct WifiApiV2 {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub bssid: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub freq: Option<i64>,
+    pub freq: Option<u32>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub quality: Option<WifiQualityApiV2>,
 }
